Create README.md
This projects is based on day today lifetstyle in which sometime we forget some important work that has to be done on some specific day. With the help of this project you can add your work list based on your workplace or work type and whenever you visit the page it will highlight all your pending work and after the work done you can easily remove it from the list and can also add very simply by just clicking a button.